+++ src/tools/wasm-fuzz-types.cpp
@@ -559,26 +559,7 @@ int main(int argc, const char* argv[]) {
               Options::Arguments::Zero,
               [&](Options*, const std::string& arg) { verbose = true; });
 
-  TypeSystem system = TypeSystem::Isorecursive;
-  options.add(
-    "--nominal",
-    "",
-    "Use the nominal type system",
-    WasmFuzzTypesOption,
-    Options::Arguments::Zero,
-    [&](Options*, const std::string& arg) { system = TypeSystem::Nominal; });
-  options.add("--hybrid",
-              "",
-              "Use the isorecursive hybrid type system (default)",
-              WasmFuzzTypesOption,
-              Options::Arguments::Zero,
-              [&](Options*, const std::string& arg) {
-                system = TypeSystem::Isorecursive;
-              });
-
   options.parse(argc, argv);
-
-  setTypeSystem(system);
 
   Fuzzer fuzzer{verbose};
   if (seed) {